refactor: [EL-4974] Extract ResourceVersionInfo component and fix copy button styling
